# public/api.yaml
openapi: 3.0.0
paths:
...
  /api/commentaires/createCommentaire:
    post:
      tags:
        - Commentaires
      summary: 'Add a comment to a game if you are an adherent'
      operationId: 5c08e5c21bad24a03197ab7608aa268f
      responses:
        '200':
          description: 'Comment created successfully'
        '422':
          description: 'Error in comment creation'
  /api/commentaires/updateCommentaire/id:
    put:
      tags:
        - Commentaires
      summary: "Update the game's comment by his id if you are the owner of the comment or if you are a comment moderator"
      operationId: ae139675bb8f0e093eba27e4805bf3bf
      responses:
        '200':
          description: 'Comment updated successfully'
        '422':
          description: 'Error in comment update'
  /api/commentaires/deleteCommentaire/id:
    delete:
      tags:
        - Commentaires
      summary: 'Delete the comment by his id if you are the owner of the comment or if you are a comment moderator'
      operationId: 9d61ed1556e4414b66367140dba0139d
      responses:
        '200':
          description: 'Comment successfully deleted'
        '422':
          description: 'Error in comment delete'
...
  /api/jeu/createJeu:
    post:
      tags:
        - Jeu
      summary: 'Add a new game if you are at least a premium adherent'
      operationId: d3b1f03e3ca9695ff00a78f70ab7e504
      responses:
        '200':
          description: 'Game created successfully!'
        '422':
          description: 'Error game create'
  /api/jeu/updateJeu/id:
    put:
      tags:
        - Jeu
      summary: 'Update the game by his id if you are at least a premium adherent'
      operationId: 070960c7e1525d46c1da13f003bd4fb3
      responses:
        '200':
          description: 'Game updated successfully!'
        '422':
          description: 'Error game update'
  /api/jeu/updateUrl/id:
    put:
      tags:
        - Jeu
      summary: "Update the game's URL by his id if you are at least a premium adherent"
      operationId: adcab4cd0db90f9361ba2e8d26a06794
      responses:
        '200':
          description: 'Game url media updated successfully!'
        '422':
          description: 'Error Game url media'
  /api/jeu/createAchat:
    post:
      tags:
        - Jeu
      summary: 'Add a purchase of a game if you are at least a premium adherent'
      operationId: 81157d1abb2b2299475020884a0e22d5
      responses:
        '200':
          description: 'Purchase created successfully!'
        '422':
          description: 'Error purchase create'
  /api/jeu/deleteAchat:
    delete:
      tags:
        - Jeu
      summary: 'Delete a purchase of a game by his id if you are at least a premium adherent or the owner of the purchase'
      operationId: 8158521432244c36bb447d44facbefda
      responses:
        '200':
          description: 'Purchase deleted successfully!'
        '422':
          description: 'Error purchase delete'
...